TASK1
Hello everyone,

This is a simple Number Guessing Game in C++. I included <iostream>, <cstdlib>, and <ctime> to handle input/output, random number generation, and seeding. Inside main(), I used srand(time(0)) to ensure a different random number each time, and rand() % 100 + 1 to pick a number between 1 and 100. The user is prompted to guess, and a do-while loop checks if the guess is too high, too low, or correct. It repeats until the correct number is guessed. This project demonstrates loops, conditionals, and basic randomization in a fun and interactive way.